+++ MBS_Now/MBS Now/HomeViewController.h
@@ -2,8 +2,8 @@
 //  HomeViewController.h
 //  MBS Now
 //
-//  Created by gdyer on 1/10/13.
-//  Copyright (c) 2013 DevelopMBS. All rights reserved.
+//  Created by Graham Dyer on 1/10/13.
+//  Copyright (c) 2013 MBS Now. All rights reserved.
 //
 
 #import <UIKit/UIKit.h>
@@ -26,7 +26,7 @@
     NSString *login;
 
     // countdown
-    int days;
+    NSInteger days;
     UIImage *bImage;
     NSDateComponents *components;
     NSString *messagePart;
